Move tests into single go test file
