--- backend/handlers/math/outcomes/dbpm/1_divideupmarketpoolsharesdbpm_test.go
@@ -1,16 +1,0 @@
-package dbpm
-
-import (
-	"testing"
-)
-
-func TestDivideUpMarketPoolSharesDBPM(t *testing.T) {
-	for _, tc := range TestCases {
-		t.Run(tc.Name, func(t *testing.T) {
-			yes, no := DivideUpMarketPoolSharesDBPM(tc.Bets, tc.ProbabilityChanges)
-			if yes != tc.S_YES || no != tc.S_NO {
-				t.Errorf("%s: expected (%d, %d), got (%d, %d)", tc.Name, tc.S_YES, tc.S_NO, yes, no)
-			}
-		})
-	}
-}
