--- backend/handlers/math/outcomes/dbpm/5_adjustpayoutsfromnewest_test.go
@@ -1,19 +1,0 @@
-package dbpm
-
-import (
-	"testing"
-)
-
-func TestAdjustPayoutsFromNewest(t *testing.T) {
-
-	for _, tc := range TestCases {
-		t.Run(tc.Name, func(t *testing.T) {
-			adjustedPayouts := AdjustPayoutsFromNewest(tc.Bets, tc.ScaledPayouts)
-			for i, payout := range adjustedPayouts {
-				if payout != tc.AdjustedScaledPayouts[i] {
-					t.Errorf("Test %s failed at index %d: expected payout %d, got %d", tc.Name, i, tc.AdjustedScaledPayouts[i], payout)
-				}
-			}
-		})
-	}
-}
