+++ backend/handlers/math/outcomes/dbpm/dbpm_marketshares_test.go
@@ -1,0 +1,129 @@
+package dbpm
+
+import (
+	"math"
+	"reflect"
+	"testing"
+)
+
+func TestAdjustPayoutsFromNewest(t *testing.T) {
+	for _, tc := range TestCases {
+		t.Run(tc.Name, func(t *testing.T) {
+			adjustedPayouts := AdjustPayoutsFromNewest(tc.Bets, tc.ScaledPayouts)
+			for i, payout := range adjustedPayouts {
+				if payout != tc.AdjustedScaledPayouts[i] {
+					t.Errorf("Test %s failed at index %d: expected payout %d, got %d", tc.Name, i, tc.AdjustedScaledPayouts[i], payout)
+				}
+			}
+		})
+	}
+}
+
+func TestAggregateUserPayoutsDBPM(t *testing.T) {
+	for _, tc := range TestCases {
+		t.Run(tc.Name, func(t *testing.T) {
+			result := AggregateUserPayoutsDBPM(tc.Bets, tc.AdjustedScaledPayouts)
+
+			// Create a map for expected results for easy lookup
+			expectedResults := make(map[string]MarketPosition)
+			for _, pos := range tc.AggregatedPositions {
+				expectedResults[pos.Username] = pos
+			}
+
+			// Create a map from the results for comparison
+			resultsMap := make(map[string]MarketPosition)
+			for _, pos := range result {
+				resultsMap[pos.Username] = pos
+			}
+
+			// Check if the results match expected results for each user
+			for username, expectedPos := range expectedResults {
+				resultPos, ok := resultsMap[username]
+				if !ok {
+					t.Errorf("Test %s failed: missing position for username %s", tc.Name, username)
+					continue
+				}
+				if resultPos.YesSharesOwned != expectedPos.YesSharesOwned || resultPos.NoSharesOwned != expectedPos.NoSharesOwned {
+					t.Errorf("Test %s failed for %s: expected %+v, got %+v", tc.Name, username, expectedPos, resultPos)
+				}
+			}
+
+			// Check for any unexpected extra users in the results
+			for username := range resultsMap {
+				if _, ok := expectedResults[username]; !ok {
+					t.Errorf("Test %s failed: unexpected position for username %s", tc.Name, username)
+				}
+			}
+		})
+	}
+}
+
+func TestCalculateCoursePayoutsDBPM(t *testing.T) {
+	for _, tc := range TestCases {
+		t.Run(tc.Name, func(t *testing.T) {
+			actualPayouts := CalculateCoursePayoutsDBPM(tc.Bets, tc.ProbabilityChanges)
+			if len(actualPayouts) != len(tc.CoursePayouts) {
+				t.Fatalf("expected %d payouts, got %d", len(tc.CoursePayouts), len(actualPayouts))
+			}
+			for i, payout := range actualPayouts {
+				if payout.Payout != tc.CoursePayouts[i].Payout || payout.Outcome != tc.CoursePayouts[i].Outcome {
+					t.Errorf("Test %s failed at index %d: expected %+v, got %+v", tc.Name, i, tc.CoursePayouts[i], payout)
+				}
+			}
+		})
+	}
+}
+
+func TestCalculateNormalizationFactorsDBPM(t *testing.T) {
+	for _, tc := range TestCases {
+		t.Run(tc.Name, func(t *testing.T) {
+			roundedF_YES := math.Round(tc.F_YES*1000000) / 1000000
+			roundedF_NO := math.Round(tc.F_NO*1000000) / 1000000
+
+			if roundedF_YES != tc.ExpectedF_YES || roundedF_NO != tc.ExpectedF_NO {
+				t.Errorf("Test %s failed: expected F_YES=%f, F_NO=%f, got F_YES=%f, F_NO=%f",
+					tc.Name, tc.ExpectedF_YES, tc.ExpectedF_NO, roundedF_YES, roundedF_NO)
+			}
+		})
+	}
+}
+
+func TestCalculateScaledPayoutsDBPM(t *testing.T) {
+	for _, tc := range TestCases {
+		t.Run(tc.Name, func(t *testing.T) {
+			actualPayouts := CalculateScaledPayoutsDBPM(tc.Bets, tc.CoursePayouts, tc.F_YES, tc.F_NO)
+			if len(actualPayouts) != len(tc.ScaledPayouts) {
+				t.Fatalf("Test %s failed: expected %d payouts, got %d", tc.Name, len(tc.ScaledPayouts), len(actualPayouts))
+			}
+			for i, payout := range actualPayouts {
+				if payout != tc.ScaledPayouts[i] {
+					t.Errorf("Test %s failed at index %d: expected payout %d, got %d", tc.Name, i, tc.ScaledPayouts[i], payout)
+				}
+			}
+		})
+	}
+}
+
+func TestDivideUpMarketPoolSharesDBPM(t *testing.T) {
+	for _, tc := range TestCases {
+		t.Run(tc.Name, func(t *testing.T) {
+			yes, no := DivideUpMarketPoolSharesDBPM(tc.Bets, tc.ProbabilityChanges)
+			if yes != tc.S_YES || no != tc.S_NO {
+				t.Errorf("%s: expected (%d, %d), got (%d, %d)", tc.Name, tc.S_YES, tc.S_NO, yes, no)
+			}
+		})
+	}
+}
+
+func TestNetAggregateMarketPositions(t *testing.T) {
+	for _, tc := range TestCases {
+		t.Run(tc.Name, func(t *testing.T) {
+			actual := NetAggregateMarketPositions(tc.AggregatedPositions)
+			expected := tc.NetPositions
+
+			if !reflect.DeepEqual(actual, expected) {
+				t.Errorf("Failed %s: expected %+v, got %+v", tc.Name, expected, actual)
+			}
+		})
+	}
+}
